directly add navigation and user code to sidebar
