+++ app/_components/organisms/index.ts
@@ -1,8 +1,6 @@
 import Wrapper from "./Wrapper";
 import Sidebar from "./Sidebar";
 import Footer from "./Footer";
-import Navigation from "./Navigation";
-import User from "./User";
 import Header from "./Header";
 import Content from "./Content";
 
@@ -10,8 +8,6 @@ const Organism = {
   Wrapper,
   Sidebar,
   Footer,
-  Navigation,
-  User,
   Header,
   Content,
 }
